refactor(tplan): isolate identity and pure execution metrics (#172)
