+++ skills/tplan/scripts/tplan_errors.py
@@ -1,0 +1,4 @@
+"""Single exception identity for TPlan runtime modules and public CLI imports."""
+
+class TplanError(ValueError):
+    """Runtime contract violation."""
